AI # PursueTargetState - 목표 추적 - 공격 사거리내에 타겟이 들어오면 Combat Stance State가 된다 - 타겟이 공격사거리 밖으로 나가면 Pursue Target State 유지 - 적 오브젝트의 상태에 따라 회전
# CombatStanceState
- 공격 사거리 확인
- 공격 대상 주위를 배회
- 공격 사거리 내에 들어가면 Attack State가 된다
- 공격후 대기상태라면 Combat Stance State로 돌아오고 타겟 주위를 배회
- 타겟이 공격사거리 밖으로 도망가면 Pursue Target State가 된다

# AttackState
- attack score 값에 따라 수행가능한 공격중 하나를 선택한다
- 선택된 공격이 각도나 거리때문에 불가능하다면 새로운 공격을 선택
- 공격이 가능하다면 이동을 멈춘 후 공격
- 공격 딜레이 시간을 세팅
- Combat Stance State로 돌아간다

# EnemyAnimatorManager
- Rigidbody를 EnemyManager에서 호출하도록 변경했기에 EnemyManager 인스턴스 호출

# EnemyLocomotionManager + EnemyManager
- 상태를 사용하기위한 여러 변경
